control if bring up manager web module



# message auto-generated for no-merge-commit merge:
!2874 merge control_up_manager_web into dev/enterprise_kub

control if bring up manager web module

Created-by: chen-hui-zhe-xi
Commit-by: chen hui
Merged-by: cyz95
Description: <!--  Thanks for sending a pull request!  Here are some tips for you:

1) If this is your first time, please read our contributor guidelines: https://gitcode.com/openJiuwen/community/blob/master/CONTRIBUTING.md

2) If you want to contribute your code but don't know who will review and merge, please add label `openjiuwen-assistant` to the pull request, we will find and do it as soon as possible.
-->

**What type of PR is this?**
<!-- 
选择下面一种标签替换下方 `/kind <label>`，可选标签类型有：
- /kind bug 
- /kind task
- /kind feature
- /kind refactor
- /kind clean_code
如PR描述不符合规范，修改PR描述后需要/check-pr重新检查PR规范。
-->
/kind bug


**Self-checklist**:（**请自检，在[ ]内打上x，我们将检视你的完成情况，否则会导致pr无法合入**）

+ - [x] **设计**：PR对应的方案是否已经经过Maintainer评审，方案检视意见是否均已答复并完成方案修改
+ - [x] **测试**：PR中的代码是否已有UT/ST测试用例进行充分的覆盖，新增测试用例是否随本PR一并上库或已经上库
+ - [x] **验证**：PR描述信息中是否已包含对该PR对应的Feature、Refactor、Bugfix的预期目标达成情况的详细验证结果描述
+ - [x] **接口**：是否涉及对外接口变更，相应变更已得到接口评审组织的通过，API对应的注释信息已经刷新正确
+ - [x] **文档**：是否涉及官网文档修改，如果涉及请及时提交资料到Doc仓

<!-- **Special notes for your reviewers**: -->
<!-- + - [ ] 是否导致无法前向兼容 -->
<!-- + - [ ] 是否涉及依赖的三方库变更 -->


See merge request: openJiuwen/jiuwenswarm!2874

+++ deploy/manager_handler.sh
@@ -6,34 +6,40 @@ deploy_manager() {
     local manager_server_name="${DEPLOY_VARS["MANAGER_SERVER_NAME"]}"
     local manager_server_template_file="${CONFIG["MANAGER_SERVER_TEMPLATE_FILE"]}"
     local manager_server_file="${CONFIG["MANAGER_SERVER_FILE"]}"
-    local manager_web_name="${DEPLOY_VARS["MANAGER_WEB_NAME"]}"
-    local manager_web_template_file="${CONFIG["MANAGER_WEB_TEMPLATE_FILE"]}"
-    local manager_web_file="${CONFIG["MANAGER_WEB_FILE"]}"
+    local is_up_web="${DEPLOY_VARS["IS_UP_MANAGER_WEB"]}"
 
     ensure_available_port "MANAGER_SERVER_NODE_PORT"
     render_config_template "${manager_server_template_file}" "${manager_server_file}" "DEPLOY_VARS"
     enable_dev_mode_if_needed ${manager_server_file}
     exec_cmd kubectl apply -f ${manager_server_file}
     wait_k8s_resource_ready "deployment" "${manager_server_name}" "${namespace}"
-
-    ensure_available_port "MANAGER_WEB_NODE_PORT"
-    render_config_template "${manager_web_template_file}" "${manager_web_file}" "DEPLOY_VARS"
-    exec_cmd kubectl apply -f ${manager_web_file}
-    wait_k8s_resource_ready "deployment" "${manager_web_name}" "${namespace}"
-
     success "MANAGER_SERVER_NODE_PORT: ${DEPLOY_VARS["MANAGER_SERVER_NODE_PORT"]}"
-    success "MANAGER_WEB_NODE_PORT: ${DEPLOY_VARS["MANAGER_WEB_NODE_PORT"]}"
+
+    if [ "${is_up_web}" == "true" ]; then
+        local manager_web_name="${DEPLOY_VARS["MANAGER_WEB_NAME"]}"
+        local manager_web_template_file="${CONFIG["MANAGER_WEB_TEMPLATE_FILE"]}"
+        local manager_web_file="${CONFIG["MANAGER_WEB_FILE"]}"
+
+        ensure_available_port "MANAGER_WEB_NODE_PORT"
+        render_config_template "${manager_web_template_file}" "${manager_web_file}" "DEPLOY_VARS"
+        exec_cmd kubectl apply -f ${manager_web_file}
+        wait_k8s_resource_ready "deployment" "${manager_web_name}" "${namespace}"
+        success "MANAGER_WEB_NODE_PORT: ${DEPLOY_VARS["MANAGER_WEB_NODE_PORT"]}"
+    fi
 }
 
 uninstall_manager() {
     local namespace="${DEPLOY_VARS["NAMESPACE"]}"
     local manager_server_name="${DEPLOY_VARS["MANAGER_SERVER_NAME"]}"
     local manager_server_file="${CONFIG["MANAGER_SERVER_FILE"]}"
-    local manager_web_name="${DEPLOY_VARS["MANAGER_WEB_NAME"]}"
-    local manager_web_file="${CONFIG["MANAGER_WEB_FILE"]}"
+    local is_up_web="${DEPLOY_VARS["IS_UP_MANAGER_WEB"]}"
 
-    exec_cmd kubectl delete -f ${manager_web_file} --ignore-not-found=true
-    wait_pod_terminated "${manager_web_name}" "${namespace}"
+    if [ "${is_up_web}" == "true" ]; then
+        local manager_web_name="${DEPLOY_VARS["MANAGER_WEB_NAME"]}"
+        local manager_web_file="${CONFIG["MANAGER_WEB_FILE"]}"
+        exec_cmd kubectl delete -f ${manager_web_file} --ignore-not-found=true
+        wait_pod_terminated "${manager_web_name}" "${namespace}"
+    fi
     exec_cmd kubectl delete -f ${manager_server_file} --ignore-not-found=true
     wait_pod_terminated "${manager_server_name}" "${namespace}"
 }
